backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
asyncpg==0.30.0
greenlet==3.2.1
strawberry-graphql[fastapi]==0.273.0
pydantic-settings==2.9.1
python-slugify==8.0.4
PyJWT==2.9.0
passlib==1.7.4
httpx==0.28.1
beautifulsoup4==4.12.3
reportlab==4.2.2
vaderSentiment==3.3.2
